* Move models * Extract dialogue history out of Quest model
