+++ common/src/main/java/com/wynntils/core/components/Models.java
@@ -10,7 +10,9 @@ import com.wynntils.models.abilities.ShamanMaskModel;
 import com.wynntils.models.abilities.ShamanTotemModel;
 import com.wynntils.models.abilitytree.AbilityTreeModel;
 import com.wynntils.models.activities.ActivityModel;
-import com.wynntils.models.activities.CaveModel;
+import com.wynntils.models.activities.caves.CaveModel;
+import com.wynntils.models.activities.discoveries.DiscoveryModel;
+import com.wynntils.models.activities.quests.QuestModel;
 import com.wynntils.models.character.CharacterModel;
 import com.wynntils.models.character.CharacterSelectionModel;
 import com.wynntils.models.characterstats.CharacterStatsModel;
@@ -20,7 +22,6 @@ import com.wynntils.models.containers.LootChestModel;
 import com.wynntils.models.containers.PlayerInventoryModel;
 import com.wynntils.models.cosmetics.CosmeticsModel;
 import com.wynntils.models.damage.DamageModel;
-import com.wynntils.models.discoveries.DiscoveryModel;
 import com.wynntils.models.elements.ElementModel;
 import com.wynntils.models.emeralds.EmeraldModel;
 import com.wynntils.models.favorites.FavoritesModel;
@@ -42,7 +43,6 @@ import com.wynntils.models.players.PartyModel;
 import com.wynntils.models.players.PlayerModel;
 import com.wynntils.models.players.hades.HadesModel;
 import com.wynntils.models.profession.ProfessionModel;
-import com.wynntils.models.quests.QuestModel;
 import com.wynntils.models.rewards.RewardsModel;
 import com.wynntils.models.seaskipper.SeaskipperModel;
 import com.wynntils.models.spells.SpellModel;
