+++ common/src/main/java/com/wynntils/models/activities/discoveries/DiscoveryModel.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.discoveries;
+package com.wynntils.models.activities.discoveries;
 
 import com.google.common.reflect.TypeToken;
 import com.wynntils.core.WynntilsMod;
@@ -14,14 +14,14 @@ import com.wynntils.core.net.Download;
 import com.wynntils.core.net.UrlId;
 import com.wynntils.core.text.StyledText;
 import com.wynntils.models.activities.event.ActivityUpdatedEvent;
+import com.wynntils.models.activities.profile.DiscoveryProfile;
+import com.wynntils.models.activities.quests.QuestModel;
 import com.wynntils.models.activities.type.ActivityInfo;
 import com.wynntils.models.activities.type.ActivitySortOrder;
 import com.wynntils.models.activities.type.ActivityType;
+import com.wynntils.models.activities.type.DiscoveryType;
 import com.wynntils.models.characterstats.CombatXpModel;
-import com.wynntils.models.discoveries.profile.DiscoveryProfile;
-import com.wynntils.models.discoveries.type.DiscoveryType;
 import com.wynntils.models.map.CompassModel;
-import com.wynntils.models.quests.QuestModel;
 import com.wynntils.models.territories.TerritoryModel;
 import com.wynntils.models.territories.profile.TerritoryProfile;
 import com.wynntils.models.worlds.event.WorldStateEvent;
