+++ common/src/main/java/com/wynntils/models/activities/event/DialogueHistoryReloadedEvent.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.quests.event;
+package com.wynntils.models.activities.event;
 
 import net.minecraftforge.eventbus.api.Event;
