+++ common/src/main/java/com/wynntils/models/activities/profile/DiscoveryProfile.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.discoveries.profile;
+package com.wynntils.models.activities.profile;
 
 import java.util.ArrayList;
 import java.util.List;
