+++ common/src/main/java/com/wynntils/models/activities/type/QuestStatus.java
@@ -2,9 +2,8 @@
  * Copyright © Wynntils 2022.
  * This file is released under AGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.models.quests.type;
+package com.wynntils.models.activities.type;
 
-import com.wynntils.models.activities.type.ActivityStatus;
 import com.wynntils.screens.activities.WynntilsQuestBookScreen;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.Component;
