+++ common/src/main/java/com/wynntils/screens/activities/WynntilsDiscoveriesScreen.java
@@ -8,9 +8,9 @@ import com.mojang.blaze3d.vertex.PoseStack;
 import com.wynntils.core.WynntilsMod;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.text.StyledText;
+import com.wynntils.models.activities.discoveries.DiscoveryInfo;
 import com.wynntils.models.activities.event.ActivityUpdatedEvent;
 import com.wynntils.models.activities.type.ActivitySortOrder;
-import com.wynntils.models.discoveries.DiscoveryInfo;
 import com.wynntils.screens.activities.widgets.DiscoveryButton;
 import com.wynntils.screens.activities.widgets.DiscoveryProgressButton;
 import com.wynntils.screens.base.WynntilsListScreen;
