+++ common/src/main/java/com/wynntils/screens/activities/WynntilsQuestBookScreen.java
@@ -11,9 +11,9 @@ import com.wynntils.core.text.StyledText;
 import com.wynntils.mc.event.MenuEvent;
 import com.wynntils.models.activities.event.ActivityTrackerUpdatedEvent;
 import com.wynntils.models.activities.event.ActivityUpdatedEvent;
+import com.wynntils.models.activities.quests.QuestInfo;
 import com.wynntils.models.activities.type.ActivitySortOrder;
-import com.wynntils.models.quests.QuestInfo;
-import com.wynntils.models.quests.type.QuestStatus;
+import com.wynntils.models.activities.type.QuestStatus;
 import com.wynntils.screens.activities.widgets.DialogueHistoryButton;
 import com.wynntils.screens.activities.widgets.QuestButton;
 import com.wynntils.screens.activities.widgets.QuestInfoButton;
