+++ common/src/main/java/com/wynntils/screens/activities/widgets/DiscoveryButton.java
@@ -7,9 +7,9 @@ package com.wynntils.screens.activities.widgets;
 import com.mojang.blaze3d.vertex.PoseStack;
 import com.wynntils.core.components.Models;
 import com.wynntils.core.text.StyledText;
+import com.wynntils.models.activities.discoveries.DiscoveryInfo;
 import com.wynntils.models.activities.type.ActivitySortOrder;
-import com.wynntils.models.discoveries.DiscoveryInfo;
-import com.wynntils.models.discoveries.type.DiscoveryType;
+import com.wynntils.models.activities.type.DiscoveryType;
 import com.wynntils.screens.base.TooltipProvider;
 import com.wynntils.screens.base.widgets.WynntilsButton;
 import com.wynntils.utils.colors.CommonColors;
